chore: Reuse existing `ImageAssert` from Xamarin UI Tests
