+++ src/Uno.UI.RuntimeTests/Helpers/RawBitmap.cs
